requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
apprise==1.9.1
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
colorama==0.4.6
coolname==2.2.0
croniter==5.0.1
cryptography==44.0.0
dateparser==1.2.0
Deprecated==1.2.15
docker==7.1.0
exceptiongroup==1.2.2
fastapi==0.115.6
fsspec==2024.12.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.1
oauthlib==3.2.2
opentelemetry-api==1.29.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
prefect==3.1.9
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.5.3
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
websockets==13.1
wrapt==1.17.0
zipp==3.21.0
